feat: Orb 잔재 추가
그냥 360도 총알이 없어서 다른거 구현하기로 함.
근데 좀.. 아까워서 일단 둔다.

+++ CupHead/MainGame.cpp
@@ -110,7 +110,7 @@ void MainGame::Render()
 
 	backGround->Render(hBackBufferDC);
 
-	wsprintf(szText, TEXT("Mouse X : %d, Y : %d"), mousePos.x, mousePos.y);
+	wsprintf(szText, TEXT("Mouse X : %d, Y : %d"), (int)mousePos.x, (int)mousePos.y);
 	TextOut(hBackBufferDC, 20, 60, szText, (int)wcslen(szText));
